created tests to check if endpoints returns ID
