--- src/test/java/com/develhope/spring/deals/models/RentalTest.java
@@ -1,49 +1,0 @@
-package com.develhope.spring.deals.models;
-import com.develhope.spring.vehicles.models.Vehicle;
-import org.junit.jupiter.api.Test;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import static org.junit.jupiter.api.Assertions.*;
-
-public class RentalTest {
-
-    private Rental rental;
-    private Vehicle vehicle;
-
-//    @Test
-//    public void testSetPaid() {
-//        rental.setPaid(true);
-//        assertTrue(rental.isPaid());
-//    }
-//
-//    @Test
-//    public void testSetStartDate() {
-//        LocalDate newStartDate = LocalDate.now().plusDays(1);
-//        rental.setStartDate(newStartDate);
-//        assertEquals(newStartDate, rental.getStartDate());
-//    }
-//
-//    @Test
-//    public void testSetEndDate() {
-//        LocalDate newEndDate = LocalDate.now().plusDays(10);
-//        rental.setEndDate(newEndDate);
-//        assertEquals(newEndDate, rental.getEndDate());
-//    }
-//
-//    @Test
-//    public void testSetDailyCost() {
-//        BigDecimal newDailyCost = new BigDecimal("150.00");
-//        rental.setDailyCost(newDailyCost);
-//        assertEquals(newDailyCost, rental.getDailyCost());
-//    }
-//
-//    @Test
-//    public void testSetTotalCost() {
-//        BigDecimal newTotalCost = new BigDecimal("600.00");
-//        rental.setTotalCost(newTotalCost);
-//        assertEquals(newTotalCost, rental.getTotalCost());
-//    }
-
-
-}
