--- src/test/java/com/develhope/spring/users/UserTest.java
@@ -1,4 +1,0 @@
-package com.develhope.spring.users;
-
-public class UserTest {
-}
