--- src/test/java/com/develhope/spring/vehicles/entities/VehicleTests.java
@@ -1,5 +1,0 @@
-package com.develhope.spring.vehicles.entities;
-
-public class VehicleTests {
-
-}
